# docs/API_SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: ClearDeed API
  version: 1.0.0
  description: End-to-end verified real estate & investment execution platform
  contact:
    name: ClearDeed Advisors LLP
servers:
  - url: https://api.cleardeed.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development
paths:
...
  /properties:
    get:
      summary: List verified properties (buyer view)
      tags:
        - Properties
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: ['land', 'individual_house', 'commercial', 'agriculture']
        - name: city
          in: query
          schema:
            type: string
        - name: min_price
          in: query
          schema:
            type: number
        - name: max_price
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of verified properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PropertyCard'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: List properties for seller upload
      tags:
        - Properties
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                  enum: ['land', 'individual_house', 'commercial', 'agriculture']
                title:
                  type: string
                location:
                  type: string
                city:
                  type: string
                pincode:
                  type: string
                price:
                  type: number
                area:
                  type: number
                area_unit:
                  type: string
                  enum: ['sqft', 'sqm']
                ownership_status:
                  type: string
                description:
                  type: string
              required:
                - category
                - title
                - location
                - city
                - price
                - area
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
  /properties/{property_id}:
    get:
      summary: Get property details with images, documents, verification status
      tags:
        - Properties
      security:
        - BearerAuth: []
      parameters:
        - name: property_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Property details with gallery and verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyDetail'
        '404':
          description: Property not found

    put:
      summary: Update property (seller only, if under_verification or submitted)
      tags:
        - Properties
      security:
        - BearerAuth: []
      parameters:
        - name: property_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                area:
                  type: number
      responses:
        '200':
          description: Property updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PropertyCard:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        city:
          type: string
        price:
          type: number
        area:
          type: number
        area_unit:
          type: string
        category:
          type: string
        image_url:
          type: string
        verified_badge:
          type: boolean
        status:
          type: string
    PropertyDetail:
      allOf:
        - $ref: '#/components/schemas/PropertyCard'
        - type: object
          properties:
            description:
              type: string
            ownership_status:
              type: string
            gallery:
              type: array
              items:
                type: object
                properties:
                  image_url:
                    type: string
                  image_title:
                    type: string
            documents:
              type: array
              items:
                type: object
                properties:
                  document_type:
                    type: string
                  document_name:
                    type: string
            verification_summary:
              type: string
    Property:
      type: object
      properties:
        id:
          type: integer
        seller_user_id:
          type: integer
        title:
          type: string
        location:
          type: string
        price:
          type: number
        area:
          type: number
        status:
          type: string
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer